requirements.txt
smolagents
openai
python-dotenv
streamlit
google-generativeai 